+++ app/Repository/Advertiser/MySqlStatsRepository.php
@@ -25,6 +25,7 @@ namespace Adshares\Adserver\Repository\Advertiser;
 
 use Adshares\Adserver\Exceptions\Advertiser\MissingEventsException;
 use Adshares\Adserver\Facades\DB;
+use Adshares\Adserver\Http\Requests\Filter\FilterCollection;
 use Adshares\Adserver\Models\Campaign;
 use Adshares\Adserver\Models\PaymentReport;
 use Adshares\Adserver\Utilities\DateUtils;
@@ -249,7 +250,8 @@ SQL;
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $result = $this->fetch(
             StatsRepository::TYPE_VIEW,
@@ -257,7 +259,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         return new ChartResult($result);
@@ -268,7 +271,8 @@ SQL;
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $result = $this->fetch(
             StatsRepository::TYPE_VIEW_ALL,
@@ -276,7 +280,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         return new ChartResult($result);
@@ -287,7 +292,8 @@ SQL;
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $resultViewsAll = $this->fetch(
             StatsRepository::TYPE_VIEW_ALL,
@@ -295,7 +301,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         $resultViews = $this->fetch(
@@ -304,7 +311,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         $result = [];
@@ -326,7 +334,8 @@ SQL;
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $result = $this->fetch(
             StatsRepository::TYPE_VIEW_UNIQUE,
@@ -334,7 +343,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         return new ChartResult($result);
@@ -345,7 +355,8 @@ SQL;
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $result = $this->fetch(
             StatsRepository::TYPE_CLICK,
@@ -353,7 +364,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         return new ChartResult($result);
@@ -364,7 +376,8 @@ SQL;
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $result = $this->fetch(
             StatsRepository::TYPE_CLICK_ALL,
@@ -372,7 +385,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         return new ChartResult($result);
@@ -383,7 +397,8 @@ SQL;
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $resultClicksAll = $this->fetch(
             StatsRepository::TYPE_CLICK_ALL,
@@ -391,7 +406,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         $resultClicks = $this->fetch(
@@ -400,7 +416,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         $result = [];
@@ -422,7 +439,8 @@ SQL;
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $resultSum = $this->fetch(
             StatsRepository::TYPE_SUM,
@@ -430,7 +448,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         $resultClicks = $this->fetch(
@@ -439,7 +458,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         $result = [];
@@ -461,7 +481,8 @@ SQL;
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $resultSum = $this->fetch(
             StatsRepository::TYPE_SUM,
@@ -469,7 +490,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         $resultViews = $this->fetch(
@@ -478,7 +500,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         $result = [];
@@ -500,7 +523,8 @@ SQL;
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $result = $this->fetch(
             StatsRepository::TYPE_SUM,
@@ -508,7 +532,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         return new ChartResult($result);
@@ -519,7 +544,8 @@ SQL;
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $result = $this->fetch(
             StatsRepository::TYPE_SUM_BY_PAYMENT,
@@ -527,7 +553,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         return new ChartResult($result);
@@ -538,7 +565,8 @@ SQL;
         string $resolution,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): ChartResult {
         $resultClicks = $this->fetch(
             StatsRepository::TYPE_CLICK,
@@ -546,7 +574,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         $resultViews = $this->fetch(
@@ -555,7 +584,8 @@ SQL;
             $resolution,
             $dateStart,
             $dateEnd,
-            $campaignId
+            $campaignId,
+            $filters,
         );
 
         $result = [];
@@ -576,7 +606,8 @@ SQL;
         ?string $advertiserId,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): DataCollection {
         $dateThreshold = $this->getDateThresholdForLiveData($dateStart->getTimezone());
 
@@ -588,7 +619,8 @@ SQL;
                 $advertiserId,
                 $dateStart,
                 min($dateEnd, (clone $dateThreshold)->modify('-1 second')),
-                $campaignId
+                $campaignId,
+                $filters
             );
         }
 
@@ -597,7 +629,8 @@ SQL;
                 $advertiserId,
                 max($dateStart, $dateThreshold),
                 $dateEnd,
-                $campaignId
+                $campaignId,
+                $filters,
             );
         }
 
@@ -683,7 +716,8 @@ SQL;
         ?string $advertiserId,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId = null
+        ?string $campaignId = null,
+        ?FilterCollection $filters = null,
     ): Total {
         $dateThreshold = $this->getDateThresholdForLiveData($dateStart->getTimezone());
 
@@ -697,7 +731,8 @@ SQL;
                 $advertiserId,
                 $dateStart,
                 min($dateEnd, (clone $dateThreshold)->modify('-1 second')),
-                $campaignId
+                $campaignId,
+                $filters,
             );
         }
 
@@ -706,7 +741,8 @@ SQL;
                 $advertiserId,
                 max($dateStart, $dateThreshold),
                 $dateEnd,
-                $campaignId
+                $campaignId,
+                $filters,
             );
         }
 
@@ -912,7 +948,7 @@ SQL;
         DateTime $dateStart,
         DateTime $dateEnd,
         ?string $campaignId,
-        ?string $bannerId = null
+        ?FilterCollection $filters = null,
     ): array {
         $dateTimeZone = $dateStart->getTimezone();
         $dateThreshold = $this->getDateThresholdForLiveData($dateTimeZone);
@@ -927,7 +963,7 @@ SQL;
                 $dateStart,
                 min($dateEnd, (clone $dateThreshold)->modify('-1 second')),
                 $campaignId,
-                $bannerId
+                $filters,
             );
 
             $concatenatedResult = self::concatenateDateColumns($dateTimeZone, $queryResult, $resolution);
@@ -941,7 +977,7 @@ SQL;
                 max($dateStart, $dateThreshold),
                 $dateEnd,
                 $campaignId,
-                $bannerId
+                $filters,
             );
 
             $concatenatedResultLive = self::concatenateDateColumns($dateTimeZone, $queryResultLive, $resolution);
@@ -961,7 +997,7 @@ SQL;
         DateTime $dateStart,
         DateTime $dateEnd,
         ?string $campaignId,
-        ?string $bannerId
+        ?FilterCollection $filters = null,
     ): array {
         $queryBuilder = (new MySqlAggregatedStatsQueryBuilder($type))
             ->setAdvertiserIds([$advertiserId])
@@ -971,11 +1007,12 @@ SQL;
         if ($campaignId) {
             $queryBuilder->appendCampaignIdWhereClause($campaignId);
         }
-
-        if ($bannerId) {
-            $queryBuilder->appendBannerIdWhereClause($bannerId);
-        } else {
-            $queryBuilder->appendAnyBannerId();
+        $queryBuilder->appendAnyBannerId();
+        if (null !== $filters) {
+            $queryBuilder->appendMediumWhereClause(
+                $filters->getFilterByName('medium'),
+                $filters->getFilterByName('vendor'),
+            );
         }
 
         $query = $queryBuilder->build();
@@ -990,7 +1027,7 @@ SQL;
         DateTime $dateStart,
         DateTime $dateEnd,
         ?string $campaignId,
-        ?string $bannerId
+        ?FilterCollection $filters = null,
     ): array {
         if (
             in_array($type, [
@@ -1009,9 +1046,11 @@ SQL;
         if ($campaignId) {
             $queryBuilder->appendCampaignIdWhereClause($campaignId);
         }
-
-        if ($bannerId) {
-            $queryBuilder->appendBannerIdWhereClause($bannerId);
+        if (null !== $filters) {
+            $queryBuilder->appendMediumWhereClause(
+                $filters->getFilterByName('medium'),
+                $filters->getFilterByName('vendor'),
+            );
         }
 
         $query = $queryBuilder->build();
@@ -1323,7 +1362,8 @@ SQL;
         ?string $advertiserId,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId
+        ?string $campaignId,
+        ?FilterCollection $filters = null,
     ): array {
         $queryBuilder =
             (new MySqlAggregatedStatsQueryBuilder(StatsRepository::TYPE_STATS))
@@ -1341,6 +1381,12 @@ SQL;
         } else {
             $queryBuilder->appendAnyBannerId();
         }
+        if (null !== $filters) {
+            $queryBuilder->appendMediumWhereClause(
+                $filters->getFilterByName('medium'),
+                $filters->getFilterByName('vendor'),
+            );
+        }
 
         $query = $queryBuilder->build();
 
@@ -1351,7 +1397,8 @@ SQL;
         ?string $advertiserId,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId
+        ?string $campaignId,
+        ?FilterCollection $filters = null,
     ): array {
         $queryBuilder =
             (new MySqlLiveStatsQueryBuilder(StatsRepository::TYPE_STATS))
@@ -1367,6 +1414,12 @@ SQL;
         if ($campaignId) {
             $queryBuilder->appendCampaignIdWhereClause($campaignId)->appendBannerIdGroupBy();
         }
+        if (null !== $filters) {
+            $queryBuilder->appendMediumWhereClause(
+                $filters->getFilterByName('medium'),
+                $filters->getFilterByName('vendor'),
+            );
+        }
 
         $query = $queryBuilder->build();
 
@@ -1377,7 +1430,8 @@ SQL;
         ?string $advertiserId,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId
+        ?string $campaignId,
+        ?FilterCollection $filters = null,
     ): array {
         $queryBuilder =
             (new MySqlAggregatedStatsQueryBuilder(StatsRepository::TYPE_STATS))
@@ -1390,6 +1444,12 @@ SQL;
         if ($campaignId) {
             $queryBuilder->appendCampaignIdWhereClause($campaignId)->appendCampaignIdGroupBy();
         }
+        if (null !== $filters) {
+            $queryBuilder->appendMediumWhereClause(
+                $filters->getFilterByName('medium'),
+                $filters->getFilterByName('vendor'),
+            );
+        }
         $queryBuilder->appendAnyBannerId();
 
         $query = $queryBuilder->build();
@@ -1401,7 +1461,8 @@ SQL;
         ?string $advertiserId,
         DateTime $dateStart,
         DateTime $dateEnd,
-        ?string $campaignId
+        ?string $campaignId,
+        ?FilterCollection $filters = null,
     ): array {
         $queryBuilder =
             (new MySqlLiveStatsQueryBuilder(StatsRepository::TYPE_STATS))
@@ -1413,6 +1474,12 @@ SQL;
 
         if ($campaignId) {
             $queryBuilder->appendCampaignIdWhereClause($campaignId)->appendCampaignIdGroupBy();
+        }
+        if (null !== $filters) {
+            $queryBuilder->appendMediumWhereClause(
+                $filters->getFilterByName('medium'),
+                $filters->getFilterByName('vendor'),
+            );
         }
 
         $query = $queryBuilder->build();