+++ CHANGELOG.md
@@ -8,6 +8,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.4.0] - 2023-02-09
+### Added
+- Filtering campaigns by medium and vendor
+- Fetching community fee from external service
+### Changed
+- Campaign can target any vendor of medium
+- Extract rejected domains from configuration API
+- Page "Why" appearance 
+### Fixed
+- Dynamic find with empty types or mimes
+- Banner classification list (correct medium)
+
 ## [2.3.2] - 2023-01-31
 ### Fixed
 - Calculating waiting payments
@@ -828,7 +840,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0]
 
-[Unreleased]: https://github.com/adshares/adserver/compare/v2.3.2...develop
+[Unreleased]: https://github.com/adshares/adserver/compare/v2.4.0...develop
+[2.4.0]: https://github.com/adshares/adserver/compare/v2.3.2...v2.4.0
 [2.3.2]: https://github.com/adshares/adserver/compare/v2.3.1...v2.3.2
 [2.3.1]: https://github.com/adshares/adserver/compare/v2.3.0...v2.3.1
 [2.3.0]: https://github.com/adshares/adserver/compare/v2.2.7...v2.3.0